pyproject.toml
[tool.poetry]
name = "jerry"
version = "0.2.0"
description = "Jerry is a LLM Agent that lives in your slack workspace"
packages = [
    {include="jerry", from="."}
]
authors = ["Arcade AI <dev@arcade.dev>"]

[build-system]
requires = ["poetry-core>=1.0.0,<2.0.0"]
build-backend = "poetry.core.masonry.api"


[tool.poetry.dependencies]
python = ">=3.10,<3.13"
slack-bolt = "~1.21.0"
python-dotenv = "^1.0.1"
langchain-mistralai = "~0.3.0"
langchain-community = "^0.3.0"
langchain = "^0.3.25"
langchain-mcp-adapters = "^0.1.7"
requests = "^2.31.0"
beautifulsoup4 = "^4.12.0"
github3-py = "^3.2.0"
google-api-python-client = "^2.100.0"
google-auth = "^2.23.0"
google-auth-oauthlib = "^1.1.0"
fastapi = {extras = ["standard"], version = "^0.110.3"}
langgraph = "^0.2.0"
duckduckgo-search = "^8.1.1"


[tool.poetry.group.dev.dependencies]
pytest = "^8.1.2"
pytest-cov = "^4.0.0"
mypy = "^1.5.1"
pre-commit = "^3.4.0"
tox = "^4.11.1"
pytest-asyncio = "^0.23.7"
ruff = "^0.6.0"
modal = "0.73.*"


[tool.mypy]
files = ["jerry"]
python_version = "3.10"
disallow_untyped_defs = "True"
disallow_any_unimported = "True"
no_implicit_optional = "True"
check_untyped_defs = "True"
warn_return_any = "True"
warn_unused_ignores = "True"
show_error_codes = "True"
ignore_missing_imports = "True"

[tool.pytest.ini_options]
testpaths = ["tests"]


[tool.coverage.run]
branch = true
source = ["jerry"]

[tool.coverage.report]
skip_empty = true
